Invoice don't need renderAllowanceOrCharge property
